# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "요약된 기술 블로그 포스트를 조회하고 사용자 인증/프로필을 관리하는 API입니다.",
        "title": "Tech-Letter 공개 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/posts": {
            "get": {
                "description": "List all posts with pagination and optional status/blog filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List posts for admin",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by blog ID",
                        "name": "blog_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by AI summary status",
                        "name": "status_ai_summarized",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by embedding status",
                        "name": "status_embedded",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationAdminPostDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new post and trigger AI summary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a post manually",
                "parameters": [
                    {
                        "description": "Post creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}": {
            "delete": {
                "description": "Delete a post by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}/embed": {
            "post": {
                "description": "Manually trigger vector embedding for a post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Trigger embedding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}/summarize": {
            "post": {
                "description": "Manually trigger AI summary for a post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Trigger AI summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
        "/chatbot/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "로그인된 사용자만 사용할 수 있는 챗봇 질의 API. 크레딧이 차감된다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chatbot"
                ],
                "summary": "챗봇 질의",
                "parameters": [
                    {
                        "description": "chat request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatbotChatRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatbotChatResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "402": {
                        "description": "크레딧 부족",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
        "/posts": {
            "get": {
                "description": "필터와 페이징 조건으로 요약된 기술 블로그 포스트 목록을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "포스트 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "페이지 번호 (1부터 시작)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지 크기 (최대 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "카테고리 목록 (OR 조건)",
                        "name": "categories",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "태그 목록 (OR 조건)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "블로그 ID",
                        "name": "blog_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "블로그 이름",
                        "name": "blog_name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "AI 요약 완료 여부",
                        "name": "status_ai_summarized",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationPostDTO"
                        }
                    }
                }
            }
        },
...
        "/posts/{id}": {
            "get": {
                "description": "ObjectID 기준으로 특정 포스트를 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "포스트 단건 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "포스트 ObjectID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PostDTO"
                        }
                    }
                }
            }
        },
...
        "/posts/{id}/view": {
            "post": {
                "description": "지정한 포스트의 조회 수(view_count)를 1 증가시킵니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "포스트 조회 수 증가",
                "parameters": [
                    {
                        "type": "string",
                        "description": "포스트 ObjectID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AdminAISummaryDTO": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "generated_at": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.AdminPostDTO": {
            "type": "object",
            "properties": {
                "aisummary": {
                    "$ref": "#/definitions/dto.AdminAISummaryDTO"
                },
                "blog_id": {
                    "type": "string"
                },
                "blog_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "embedding": {
                    "$ref": "#/definitions/dto.AdminPostEmbeddingDTO"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/dto.AdminPostStatusDTO"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
        "dto.AdminPostEmbeddingDTO": {
            "type": "object",
            "properties": {
                "chunk_count": {
                    "type": "integer"
                },
                "collection_name": {
                    "type": "string"
                },
                "embedded_at": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "vector_dimension": {
                    "type": "integer"
                }
            }
        },
        "dto.AdminPostStatusDTO": {
            "type": "object",
            "properties": {
                "ai_summarized": {
                    "type": "boolean"
                },
                "embedded": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.ChatbotChatRequestDTO": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string",
                    "example": "벡터 DB는 어떤 원리로 동작하나요?"
                },
                "session_id": {
                    "description": "Optional for first message (will create new session if empty, but client should create session first ideally)",
                    "type": "string"
                }
            }
        },
        "dto.ChatbotChatResponseDTO": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "consumed_credits": {
                    "type": "integer"
                },
                "remaining_credits": {
                    "type": "integer"
                }
            }
        },
        "dto.ErrorResponseDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_token"
                }
            }
        },
...
        "dto.MessageResponseDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "view count incremented successfully"
                }
            }
        },
        "dto.PaginationAdminPostDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AdminPostDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PaginationPostDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PostDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.PostDTO": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                },
                "blog_name": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_bookmarked": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer 토큰. 형식: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}